voice/Berkeley/requirements.txt
flask
openai
azure-cognitiveservices-speech
python-dotenv
websockets
asyncio
uuid
quart
aiohttp